AXON-349: Improve saga replay handling  - Conditionally allow event replay for saga manages via "allowReplay"    property on manager.  - Apply replay protection and configuration to all saga managers,    including the AsyncAnnotatedSagaManager  - Refactored SagaManager class hierarchy to have a common superclass    for all synchronous as well as asynchronous implementations  - AbstractSagaManager -> AbstractSynchronousSagaManager  - pulled up shared functionality required by all saga managers, incl.    replay protection.  - Updated spring xml bindings to configure "allowReplay" attribute to    be defined in the axon:saga-manager schema.
