+++ core/src/main/java/org/axonframework/saga/annotation/AsyncAnnotatedSagaManager.java
@@ -36,6 +36,7 @@ import org.axonframework.eventhandling.EventBus;
 import org.axonframework.eventhandling.EventProcessingMonitor;
 import org.axonframework.eventhandling.EventProcessingMonitorCollection;
 import org.axonframework.eventhandling.EventProcessingMonitorSupport;
+import org.axonframework.saga.AbstractSagaManager;
 import org.axonframework.saga.GenericSagaFactory;
 import org.axonframework.saga.SagaCreationPolicy;
 import org.axonframework.saga.SagaFactory;
@@ -69,7 +70,7 @@ import java.util.concurrent.TimeUnit;
  * @author Allard Buijze
  * @since 2.0
  */
-public class AsyncAnnotatedSagaManager implements SagaManager, Subscribable, EventProcessingMonitorSupport {
+public class AsyncAnnotatedSagaManager extends AbstractSagaManager<AbstractAnnotatedSaga> implements EventProcessingMonitorSupport {
 
     private static final WaitStrategy DEFAULT_WAIT_STRATEGY = new BlockingWaitStrategy();
     private WaitStrategy waitStrategy = DEFAULT_WAIT_STRATEGY;
@@ -77,8 +78,6 @@ public class AsyncAnnotatedSagaManager implements SagaManager, Subscribable, Eve
     private int bufferSize = DEFAULT_BUFFER_SIZE;
     private static final int DEFAULT_PROCESSOR_COUNT = 1;
     private int processorCount = DEFAULT_PROCESSOR_COUNT;
-    private final EventBus eventBus;
-    private final Class<? extends AbstractAnnotatedSaga>[] sagaTypes;
     private final ParameterResolverFactory parameterResolverFactory;
     private final SagaManagerStatus sagaManagerStatus = new SagaManagerStatus();
     private final EventProcessingMonitorCollection processingMonitors = new EventProcessingMonitorCollection();
@@ -89,7 +88,6 @@ public class AsyncAnnotatedSagaManager implements SagaManager, Subscribable, Eve
     private volatile SagaFactory sagaFactory = new GenericSagaFactory();
     private UnitOfWorkFactory unitOfWorkFactory = new DefaultUnitOfWorkFactory();
     private long startTimeout = 5000;
-    private CorrelationDataProvider<? super EventMessage> correlationDataProvider = new SimpleCorrelationDataProvider();
     private ErrorHandler errorHandler = new ProceedingErrorHandler();
 
     /**
@@ -105,9 +103,8 @@ public class AsyncAnnotatedSagaManager implements SagaManager, Subscribable, Eve
      */
     @Deprecated
     public AsyncAnnotatedSagaManager(EventBus eventBus, Class<? extends AbstractAnnotatedSaga>... sagaTypes) {
+        super(eventBus,sagaTypes);
         Assert.notNull(eventBus, "eventBus may not be null");
-        this.eventBus = eventBus;
-        this.sagaTypes = Arrays.copyOf(sagaTypes, sagaTypes.length);
         this.parameterResolverFactory = ClasspathParameterResolverFactory.forClass(
                 sagaTypes.length == 0 ? AsyncAnnotatedSagaManager.class : sagaTypes[0]);
     }
@@ -135,9 +132,8 @@ public class AsyncAnnotatedSagaManager implements SagaManager, Subscribable, Eve
      */
     public AsyncAnnotatedSagaManager(ParameterResolverFactory parameterResolverFactory,
                                      Class<? extends AbstractAnnotatedSaga>... sagaTypes) {
+        super(null, sagaTypes);
         this.parameterResolverFactory = parameterResolverFactory;
-        this.eventBus = null;
-        this.sagaTypes = Arrays.copyOf(sagaTypes, sagaTypes.length);
     }
 
     /**
@@ -358,28 +354,6 @@ public class AsyncAnnotatedSagaManager implements SagaManager, Subscribable, Eve
     public synchronized void setWaitStrategy(WaitStrategy waitStrategy) {
         Assert.state(disruptor == null, "Cannot set waitStrategy when SagaManager has started");
         this.waitStrategy = waitStrategy;
-    }
-
-    /**
-     * Sets the CorrelationDataProvider which provides the Meta Data to assign to messages generated by the Sagas that
-     * are managed by this SagaManager instance.
-     *
-     * @param correlationDataProvider the instance that provides the information to attach as correlation data
-     */
-    public synchronized void setCorrelationDataProvider(
-            CorrelationDataProvider<? super EventMessage> correlationDataProvider) {
-        this.correlationDataProvider = correlationDataProvider;
-    }
-
-    /**
-     * Sets the CorrelationDataProviders which provide the Meta Data to assign to messages generated by the Sagas that
-     * are managed by this SagaManager instance.
-     *
-     * @param correlationDataProviders a list of instances that provides the information to attach as correlation data
-     */
-    public synchronized void setCorrelationDataProviders(
-            List<? extends CorrelationDataProvider<? super EventMessage>> correlationDataProviders) {
-        this.correlationDataProvider = new MultiCorrelationDataProvider<EventMessage>(correlationDataProviders);
     }
 
     private static final class SagaProcessingEventTranslator implements EventTranslator<AsyncSagaProcessingEvent> {
